# portal.openapi.yaml
openapi: 3.1.0
info:
  title: Portal Tattoo API
  version: 1.0.0
  description: Public and administrative HTTP contract for Portal Tattoo.
servers:
  - url: /api/v1
...
security: []
paths:
...
  /admin/media:
    post:
      tags: [Admin Media]
      security: [{oidc: []}]
      operationId: registerMedia
      x-required-capability: media.review
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterMediaRequest'}
      responses:
        '201':
          description: Media registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MediaAsset'}
  /admin/media/{id}/rights:
    post:
      tags: [Admin Media]
      security: [{oidc: []}]
      operationId: setMediaRights
      x-required-capability: media.review
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rightsStatus, reason]
              properties:
                rightsStatus: {$ref: '#/components/schemas/MediaRightsStatus'}
                reason: {type: string, minLength: 3}
      responses:
        '200': {description: Rights status updated}
        '409': {$ref: '#/components/responses/ConcurrentModification'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Aggregate version, formatted as a quoted integer ETag.
      schema: {type: string, pattern: '^"[0-9]+"$'}
  responses:
...
    ConcurrentModification:
      description: Aggregate changed since client read
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Problem:
      type: object
      required: [type,title,status,code,correlationId]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        correlationId: {$ref: '#/components/schemas/Uuid'}
...
    MediaRightsStatus:
      type: string
      enum: [UNKNOWN,PENDING,PERMITTED,RESTRICTED,EXPIRED,TAKEDOWN]
    RegisterMediaRequest:
      type: object
      required: [kind,storageKey,mimeType,byteSize,sha256]
      properties:
        kind: {type: string}
        storageKey: {type: string}
        mimeType: {type: string}
        byteSize: {type: integer, minimum: 0}
        sha256: {type: string}
        sourceUrl: {type: string, format: uri}
    MediaAsset:
      type: object
      required: [id,kind,mimeType,rightsStatus,status,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        kind: {type: string}
        mimeType: {type: string}
        rightsStatus: {$ref: '#/components/schemas/MediaRightsStatus'}
        status: {type: string}
        version: {type: integer}
...